+++ klaw-channel/src/manager.rs
@@ -1,6 +1,5 @@
 use crate::{ChannelResult, ChannelRuntime, dingtalk::DingtalkChannel, telegram::TelegramChannel};
 use klaw_config::{ChannelsConfig, DingtalkConfig, TelegramConfig};
-use klaw_voice::VoiceService;
 use std::{
     collections::{BTreeMap, BTreeSet},
     fmt,
@@ -205,16 +204,7 @@ pub trait ChannelDriverFactory {
 }
 
 #[derive(Debug, Default, Clone)]
-pub struct DefaultChannelDriverFactory {
-    voice_service: Option<Arc<VoiceService>>,
-}
-
-impl DefaultChannelDriverFactory {
-    #[must_use]
-    pub fn new(voice_service: Option<Arc<VoiceService>>) -> Self {
-        Self { voice_service }
-    }
-}
+pub struct DefaultChannelDriverFactory;
 
 impl ChannelDriverFactory for DefaultChannelDriverFactory {
     fn build(
@@ -226,10 +216,7 @@ impl ChannelDriverFactory for DefaultChannelDriverFactory {
                 Ok(Box::new(DingtalkChannel::from_app_config(config.clone())?))
             }
             ChannelInstanceConfig::Telegram(config) => {
-                Ok(Box::new(TelegramChannel::from_app_config_with_voice(
-                    config.clone(),
-                    self.voice_service.clone(),
-                )?))
+                Ok(Box::new(TelegramChannel::from_app_config(config.clone())?))
             }
         }
     }